Feedback fixes to the modal and toast
