+++ scripts/scripts.js
@@ -149,11 +149,9 @@ function buildAutoBlocks(main) {
   }
 }
 
-function decorateModals(main) {
-  const ctaModalButton = main.querySelector('.ctasection p');
-  const content = () => {
-
-  };
+export function decorateModals(main) {
+  const ctaModalButton = main.querySelector('.show-modal-btn');
+  const content = () => (ctaModalButton.getAttribute('data-dialog-message') ? ctaModalButton.getAttribute('dialog-message') : '');
   // Listens to the custom modal button
   ctaModalButton?.addEventListener('click', async (e) => {
     e.preventDefault();
@@ -177,7 +175,6 @@ export function decorateMain(main) {
   buildAutoBlocks(main);
   decorateSections(main);
   decorateBlocks(main);
-  decorateModals(main);
 }
 
 /**
@@ -300,70 +297,6 @@ function loadDelayed() {
     return import('./delayed.js');
   }, 3000);
   // load anything that can be postponed to the latest here
-}
-
-/**
- * Get a cookie
- * @param cname the name of the cookie
- */
-export function getCookie(cname) {
-  let value = decodeURIComponent(
-    // eslint-disable-next-line prefer-template
-    document.cookie.replace(new RegExp('(?:(?:^|.*;)\\s*' + encodeURIComponent(cname).replace(/[\\-\\.\\+\\*]/g, '\\$&') + '\\s*\\=\\s*([^;]*).*$)|^.*$'), '$1'),
-  ) || null;
-  if (value && ((value.substring(0, 1) === '{' && value.substring(value.length - 1, value.length) === '}') || (value.substring(0, 1) === '[' && value.substring(value.length - 1, value.length) === ']'))) {
-    try {
-      value = JSON.parse(value);
-    } catch (e) {
-      return value;
-    }
-  }
-  return value;
-}
-
-/**
-* Set the content of a cookie
-* @param {string} cname The cookie name (or property)
-* @param {string} cvalue The cookie value
-* @param {number} expTime The cookie expiry time (default 30 days)
-* @param {string} path The cookie path (optional)
-*
-*/
-export function setCookie(cname, cvalue, expTime = 30 * 1000 * 60 * 60 * 24, path = '/') {
-  const today = new Date();
-  today.setTime(today.getTime() + (expTime));
-  const expires = 'expires='.concat(today.toGMTString());
-  const cookieString = cname.concat('=')
-    .concat(cvalue)
-    .concat(';')
-    .concat(expires)
-    .concat(';path=')
-    .concat(path);
-  document.cookie = cookieString; // cname + '=' + cvalue + ';' + expires + ';path=' + path;
-}
-
-/**
- * Returns the user logged in state based cookie
- */
-export function isLoggedInUser() {
-  return getCookie('rationalized_id');
-}
-
-/**
- * Returns the user authorization used for commerce API calls
- */
-export function getAuthorization() {
-  const authHeader = new Headers();
-  if (localStorage.getItem('authToken')) {
-    authHeader.append('Authorization', `Bearer ${localStorage.getItem('authToken')}`);
-  } else if (getCookie('ProfileData')) {
-    const { customer_token: apiToken } = getCookie('ProfileData');
-    authHeader.append('authentication-token', apiToken);
-  } else if (getCookie('apiToken')) {
-    const apiToken = getCookie('apiToken');
-    authHeader.append('authentication-token', apiToken);
-  }
-  return authHeader;
 }
 
 async function loadPage() {
