+++ styles/modal/modal.css
@@ -1,26 +1,3 @@
 dialog::backdrop {
   @apply inset-0 bg-black bg-opacity-50;
 }
-
-.modal-title span.icon {
-  @apply flex items-center justify-center flex-shrink-0 mx-auto bg-gray-200 rounded-full w-10 h-10 p-2;
-}
-
-.quote-textarea{
-  @apply block w-full px-1 py-2 border border-gray-300 rounded-md shadow-sm focus:border-gray-300 focus:ring-gray-300 sm:text-sm;
-}
-
-.quote{
-  @apply sm:flex-row flex-col;
-}
-
-.btn-outline-trending-brand {
-  @apply text-danaherpurple-500 bg-white border-2 border-danaherpurple-500 hover:text-white hover:bg-danaherpurple-800;
-}
-
-.quote-tip{
-  @apply flex flex-col p-4 mt-4 rounded-md bg-danaherlightblue-500 bg-opacity-10;
-}
-.border-red{
-  @apply border-red-500;
-}
