+++ styles/styles.css
@@ -1210,113 +1210,6 @@ dialog::backdrop{
     --tw-bg-opacity: 0.5;
 }
 
-.modal-title span.icon{
-    margin-left: auto;
-    margin-right: auto;
-    display: flex;
-    height: 2.5rem;
-    width: 2.5rem;
-    flex-shrink: 0;
-    align-items: center;
-    justify-content: center;
-    border-radius: 9999px;
-    --tw-bg-opacity: 1;
-    background-color: rgb(229 231 235 / var(--tw-bg-opacity));
-    padding: 0.5rem;
-}
-
-.quote-textarea{
-    display: block;
-    width: 100%;
-    border-radius: 0.375rem;
-    border-width: 1px;
-    --tw-border-opacity: 1;
-    border-color: rgb(209 213 219 / var(--tw-border-opacity));
-    padding-left: 0.25rem;
-    padding-right: 0.25rem;
-    padding-top: 0.5rem;
-    padding-bottom: 0.5rem;
-    --tw-shadow: 0 1px 2px 0 rgb(0 0 0 / 0.05);
-    --tw-shadow-colored: 0 1px 2px 0 var(--tw-shadow-color);
-    box-shadow: var(--tw-ring-offset-shadow, 0 0 #0000), var(--tw-ring-shadow, 0 0 #0000), var(--tw-shadow);
-}
-
-.quote-textarea:focus{
-    --tw-border-opacity: 1;
-    border-color: rgb(209 213 219 / var(--tw-border-opacity));
-    --tw-ring-opacity: 1;
-    --tw-ring-color: rgb(209 213 219 / var(--tw-ring-opacity));
-}
-
-@media (min-width: 640px){
-
-    .quote-textarea{
-        font-size: 0.875rem;
-        line-height: 1.25rem;
-    }
-}
-
-.quote{
-    flex-direction: column;
-}
-
-.testimonial .testimonial-footer .quote>div:nth-child(2){
-    margin-top: 0px;
-    font-size: 1rem;
-    font-weight: 500;
-    line-height: 1.5rem;
-    --tw-text-opacity: 1;
-    color: rgb(2 105 124 / var(--tw-text-opacity));
-}
-
-@media (min-width: 640px){
-
-    .quote{
-        flex-direction: row;
-    }
-}
-
-.btn-outline-trending-brand{
-    border-width: 2px;
-    --tw-border-opacity: 1;
-    border-color: rgb(117 35 255 / var(--tw-border-opacity));
-    --tw-bg-opacity: 1;
-    background-color: rgb(255 255 255 / var(--tw-bg-opacity));
-    --tw-text-opacity: 1;
-    color: rgb(117 35 255 / var(--tw-text-opacity));
-}
-
-.btn-outline-trending-brand:hover{
-    --tw-bg-opacity: 1;
-    background-color: rgb(64 0 165 / var(--tw-bg-opacity));
-    --tw-text-opacity: 1;
-    color: rgb(255 255 255 / var(--tw-text-opacity));
-}
-
-.quote-tip{
-    margin-top: 1rem;
-    display: flex;
-    flex-direction: column;
-    border-radius: 0.375rem;
-    background-color: rgb(59 199 229 / var(--tw-bg-opacity));
-    --tw-bg-opacity: 0.1;
-    padding: 1rem;
-}
-
-.testimonial .testimonial-footer .quote-tip>div:nth-child(2){
-    margin-top: 0px;
-    font-size: 1rem;
-    font-weight: 500;
-    line-height: 1.5rem;
-    --tw-text-opacity: 1;
-    color: rgb(2 105 124 / var(--tw-text-opacity));
-}
-
-.border-red{
-    --tw-border-opacity: 1;
-    border-color: rgb(239 68 68 / var(--tw-border-opacity));
-}
-
 /*
  ! tailwindcss v3.3.3 | MIT License | https://tailwindcss.com
  */
@@ -2572,6 +2465,10 @@ main .section{
     height: 5rem !important;
 }
 
+.h-10{
+    height: 2.5rem;
+}
+
 .h-12{
     height: 3rem;
 }
@@ -2655,6 +2552,10 @@ main .section{
 
 .w-1\/2{
     width: 50%;
+}
+
+.w-10{
+    width: 2.5rem;
 }
 
 .w-12{
@@ -3042,6 +2943,11 @@ main .section{
     border-color: rgb(117 35 255 / var(--tw-border-opacity));
 }
 
+.border-gray-300{
+    --tw-border-opacity: 1;
+    border-color: rgb(209 213 219 / var(--tw-border-opacity));
+}
+
 .border-gray-600{
     --tw-border-opacity: 1;
     border-color: rgb(75 85 99 / var(--tw-border-opacity));
@@ -3050,6 +2956,11 @@ main .section{
 .border-purple-200{
     --tw-border-opacity: 1;
     border-color: rgb(125 86 164 / var(--tw-border-opacity));
+}
+
+.border-red-500{
+    --tw-border-opacity: 1;
+    border-color: rgb(239 68 68 / var(--tw-border-opacity));
 }
 
 .border-transparent{
@@ -3128,6 +3039,10 @@ main .section{
     background-color: rgb(255 255 255 / var(--tw-bg-opacity));
 }
 
+.bg-opacity-10{
+    --tw-bg-opacity: 0.1;
+}
+
 .fill-white{
     fill: #fff;
 }
@@ -3161,6 +3076,11 @@ main .section{
 .\!py-1{
     padding-top: 0.25rem !important;
     padding-bottom: 0.25rem !important;
+}
+
+.px-1{
+    padding-left: 0.25rem;
+    padding-right: 0.25rem;
 }
 
 .px-2{
@@ -3566,6 +3486,12 @@ main .section{
     box-shadow: var(--tw-ring-offset-shadow, 0 0 #0000), var(--tw-ring-shadow, 0 0 #0000), var(--tw-shadow);
 }
 
+.shadow-sm{
+    --tw-shadow: 0 1px 2px 0 rgb(0 0 0 / 0.05);
+    --tw-shadow-colored: 0 1px 2px 0 var(--tw-shadow-color);
+    box-shadow: var(--tw-ring-offset-shadow, 0 0 #0000), var(--tw-ring-shadow, 0 0 #0000), var(--tw-shadow);
+}
+
 .shadow-xl{
     --tw-shadow: 0 20px 25px -5px rgb(0 0 0 / 0.1), 0 8px 10px -6px rgb(0 0 0 / 0.1);
     --tw-shadow-colored: 0 20px 25px -5px var(--tw-shadow-color), 0 8px 10px -6px var(--tw-shadow-color);
@@ -3672,6 +3598,11 @@ main .section{
     background-color: rgb(243 244 246 / var(--tw-bg-opacity));
 }
 
+.hover\:bg-danaherpurple-800:hover{
+    --tw-bg-opacity: 1;
+    background-color: rgb(64 0 165 / var(--tw-bg-opacity));
+}
+
 .hover\:bg-gray-100:hover{
     --tw-bg-opacity: 1;
     background-color: rgb(243 244 246 / var(--tw-bg-opacity));
@@ -3725,6 +3656,16 @@ main .section{
     color: rgb(255 255 255 / var(--tw-text-opacity));
 }
 
+.focus\:border-gray-300:focus{
+    --tw-border-opacity: 1;
+    border-color: rgb(209 213 219 / var(--tw-border-opacity));
+}
+
+.focus\:ring-gray-300:focus{
+    --tw-ring-opacity: 1;
+    --tw-ring-color: rgb(209 213 219 / var(--tw-ring-opacity));
+}
+
 .group:hover .group-hover\:block{
     display: block;
 }
@@ -3748,6 +3689,10 @@ main .section{
         grid-template-columns: repeat(2, minmax(0, 1fr));
     }
 
+    .sm\:flex-row{
+        flex-direction: row;
+    }
+
     .sm\:px-0{
         padding-left: 0px;
         padding-right: 0px;
@@ -3761,6 +3706,11 @@ main .section{
     .sm\:text-3xl{
         font-size: 1.875rem;
         line-height: 2.25rem;
+    }
+
+    .sm\:text-sm{
+        font-size: 0.875rem;
+        line-height: 1.25rem;
     }
 }
